Adds ifEmpty() and ifNotEmpty() methods
This is a proposal to add some common conditional execution methods (or a generic `if($condition, $callback)`). The pull request itself adds ifEmpty(..) and ifNotEmpty(..) mostly to show what I had in mind. I wanted to set up this pull request to ask if this is something desirable before going in and adding a bunch.

Here's the situation this solves in my code:
```

$pages = new Map(array(
    FilterPage::Create(),
    WalletsPage::Create(),
    WalletPage::Create(),
    CategoryPage::Create(),
    TagPage::Create(),
    StorePage::Create(),
));

$dictionary = array();
$dictionary["GetParameterValues"] = $pageParameters->Render();

$page = $pages
    ->filter(fn(Page $page, $_) => $page->GetIdentifier() == $pageParameters->Page)
    ->ifEmpty(fn ($_) => die("unknown page: $pageParameters->Page"))
    ->first();
CastToPage($page)
    ->SetDictionary($dictionary)
    ->SetConfig($config)
    ->SetPageParameters($pageParameters)
    ->SetTransactionSelections($account)
    ->Process()
    ->Render();
```